Additional choices, public polls page
