+++ index.html
@@ -23,6 +23,23 @@
   <code id="errorInfo">Nothing yet.</code><br>
   <button onclick="document.getElementById('error').style.display = 'none';">Close</button>
 </dialog>
+<dialog id="extraOptions">
+  <h1>Extras</h1>
+  <p>Here you can add some extra options to your poll.</p>
+  <div class="twoColumns">
+    <label for="allowMultipleVotes">Allow multiple votes</label>
+    <input type="checkbox" id="allowMultipleVotes" />
+  </div>
+  <div class="twoColumns">
+    <label for="allowComments">Allow comments</label>
+    <input type="checkbox" id="allowComments" />
+  </div>
+  <div class="twoColumns">
+    <label for="allowPublic">Allow public voting</label>
+    <input type="checkbox" id="allowPublic" />
+  </div>
+  <button onclick="document.getElementById('extraOptions').style.display='none'">Close</button>
+</dialog>
 
 <body>
   <div id="screen">
@@ -37,8 +54,8 @@
       <button id="makeTabBtn" class="tabBtn">
         <span>Make Poll</span>
       </button>
-      <button id="helpTabBtn" class="tabBtn">
-        <span>Help</span>
+      <button id="moreTabBtn" class="tabBtn">
+        <span>More</span>
       </button>
     </div>
     <div id="tabs">
@@ -90,7 +107,9 @@
         <br />
         <input type="text" id="option4Input" placeholder="Option 4" required maxlength="200" />
         <br />
-        <button onclick="makePoll()">Make Poll</button><br><br>
+        <button onclick="document.getElementById('extraOptions').style.display='block'">Extras</button>
+        <button onclick="makePoll()">Make Poll</button>
+        <br><br>
         <div id="shareTab">
           <hr>
           <h2>Share Poll</h2>
@@ -100,9 +119,20 @@
         </div>
         <div id="loader2" class="loader"></div>
       </div>
-      <div id="helpTab" class="tab">
-        <h2>Help</h2>
-        <iframe src="./help.html" frameborder="0" height="330"></iframe>
+      <div id="moreTab" class="tab">
+        <h2>More</h2>
+        <div class="grid">
+          <button>Public Polls</button>
+          <button>View Poll</button>
+          <button>Add-ons</button>
+          <button>Help</button>
+        </div>
+        <iframe src="./help.html" frameborder="0" height="330" id="helpTab"></iframe>
+        #<div id="publicPollsTab">
+          <h2>Public Polls</h2>
+          <p>Polls that are made publically are displayed here.</p>
+          <button id="loadPublicPollsBtn">Load Public Polls</button>
+          <div id="publicPollsList"></div>
       </div>
     </div>
     <!-- <footer>
@@ -114,5 +144,6 @@
   </div>
 </body>
 <script src="./script.js"></script>
+<script src="./tabs.js"></script>
 
 </html>
