+++ style.css
@@ -146,7 +146,8 @@ iframe{
 #accessTab input,
 #accessTab button,
 #reportPoll button,
-#error button{
+#error button,
+dialog button{
      width: 80%;
      margin: auto;
      padding: 5px;
@@ -194,6 +195,7 @@ dialog{
      width: 80%;
      max-width: 500px;
      padding: 20px;
+     min-height: 60%;
 }
 #reportPollText{
      text-align: left;
@@ -224,4 +226,22 @@ dialog{
 @keyframes spin {
   0% { transform: rotate(0deg); }
   100% { transform: rotate(360deg); }
+}
+
+.grid{
+     display: grid;
+     grid-template-columns: 50% 50%;
+     grid-template-rows: 50% 50%;
+     gap: 10px;
+     padding: 10px;
+}
+
+.grid > button{
+     width: 100%;
+     height: 100%;
+     font-family: 'Poppins';
+     background-color: black;
+     color: white;
+     border: none;
+     padding: 5px;
 }
